requirements.txt
# Upbit/Bybit Algo-Trading Backend

## Requirements
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
pyupbit>=0.2.33
pybit>=5.6.0
apscheduler>=3.10.0
cryptography>=42.0.0
python-telegram-bot>=20.0
httpx>=0.26.0
pandas>=2.0.0
ta>=0.11.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
sse-starlette>=1.8.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.0.0


